requirements.txt
djproxy
django-extensions
werkzeug
pyopenssl
